requirements.txt
typer==0.12.3
requests==2.31.0
rich==13.7.0
pyyaml==6.0.1
psutil==5.9.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
websockets==12.0

# Monitoring
prometheus-client==0.19.0

# Async
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Dev tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: AWS Bedrock 지원
# boto3>=1.34.0

# Optional: PostgreSQL
# psycopg2-binary>=2.9.9
